Build package.json include a CI pre-push check
---cypress.json & config/cypress.ci.json---

Including cypress.json seems redundant however, the CLI command `cypress run` requires cypress.json in the config and will not run without it.

On the other hand for `cypress open` the plugin is powerup first to find the correct cypress config files.

--config/lintstagedrc.yml- & package.json--
Removed yarn run flow:quick-check as lintstaged was not able to run the static checker. Thus it was included as part of HuskyJS